requirements.txt
# Core dependencies for LAS file reading and data processing
lasio>=0.30
pandas>=2.0.0
numpy>=1.24.0

# Vector database and embeddings
chromadb>=0.4.0
sentence-transformers>=2.2.0

# Optional: OpenAI embeddings (if using OpenAI instead of sentence-transformers)
# openai>=1.0.0

# Visualization (following referenced notebook pattern)
altair>=5.0.0

# RAG framework and LLM integration
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
langchain-google-genai>=1.0.0
openai>=1.0.0
google-generativeai>=0.3.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0

# Frontend
streamlit>=1.28.0

# PDF processing (for petrophysical reports if needed)
# PyPDF2>=3.0.0

